Move endpoint to webapp
The endpoint has been moved to a webapp, so that it
is fixed across projects, and can be configured by a constant.

Two new query params – xpResourcesDirPath and xpAppName – is
now used to configure the FreeMarker template loaders.

A variant of ResourceTemplateLoader has been added so that
templates contained in libraries can be loaded by the resource
URL, which consists of the appName and path in XP.